MAINT: moved HessianPrecon into its own file
